requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
google-generativeai
firebase_admin
python-multipart